Jira-ID: CRNT-26798 Fixed casting bug (Float to Double / Double to Double)
git-svn-id: http://emeafrazerg/svn/ipp/product/trunk/stardust/engine@60376 8100b5e0-4d52-466c-ae9c-bdeccbdeaf6b

+++ stardust-engine-core/src/main/java/org/eclipse/stardust/engine/core/runtime/internal/SyncCriticalitiesToDiskAction.java
@@ -136,8 +136,8 @@ public class SyncCriticalitiesToDiskAction extends Procedure
                         stmt.addBatch(MessageFormat.format(
                               updateTableSql,
                               new Object[] {
-                                    DmlManager.getSQLValue(Double.TYPE, new Float(
-                                          criticality), jdbcSession.getDBDescriptor()),
+                                    DmlManager.getSQLValue(Double.TYPE, new Double(criticality), 
+                                          jdbcSession.getDBDescriptor()),
                                     DmlManager.getSQLValue(Long.TYPE, new Long(aiOid),
                                           jdbcSession.getDBDescriptor())}));
                      }
